move SNS utils to separate file
